Add support for storing credentials in file
A file store could be used as a credential cache, which is simple and
effective. You won't be using the stored credentials blindly as there
will be a validation process after retrieving it, with a follow-up normal
authentication (if it turns out to be invalid).

It's now at user's choice to use the old keyring based storage ("keyring")
or this newly added file backend ("file"). See "-s".

The new design uses interface to keep code flexible.